Modifico Servicios y comento paginacion

+++ app/src/main/java/ar/edu/unlam/mobile/scaffolding/data/network/api/dto/FeedResponseDto.kt
@@ -18,7 +18,7 @@ data class FeedResponseDto
         @Json(name = "avatar_url")
         val avatar_url: String,
         @Json(name = "likes")
-        val likes: Int,
+        val likes: Long,
         @Json(name = "liked")
         val liked: Boolean,
         @Json(name = "date")
@@ -26,6 +26,7 @@ data class FeedResponseDto
     )
 
 fun FeedResponseDto.toDomain(): Tuit =
+
     Tuit(
         id = this.id,
         message = this.message,
